chore: add tests on paginated relationships
Add tests to verify that relationships can be queried on the results of
GraphQL queries and mutations.

Querying paginated relationships on the result of deletions doesn't
correctly work yet.
Indeed, Ash currently queries relationships using a lateral join but
after the resource deletion has happened, so it looks like nothing is
related.

Signed-off-by: Davide Briani <davide@briani.dev>