fix: change default playground interface (#319)
Previously, the default playground interface was set to `:playground`
which uses graphql-playground. However, this uses a deprecated
JavaScript API which means that it does not work on modern versions of
chrome. It also seems that the library is no longer maintained, so the
chances of this being fixed in a future release are slim.

The result of this, is that tooltips do not disappear, making the
application seem broken. Whilst this isn't a direct issue of this
library, I think it gives a bad impression when things are broken out of
the box.

This commit changes the generated interface to graphiql (using the
`:simple` interface). There is also an `:advanced` interface option, but
that also points to an unmaintained library which didn't work out of the
box.

https://github.com/graphql/graphql-playground/issues/1429